sep1/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
Jinja2
db-sqlite3
python-multipart